pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"

requires = [ "setuptools>=80.1" ]

[project]
name = "stash"
description = "StaSh for PC"
readme = "README.md"
authors = [
  { name = "https://github.com/ywangd" },
]
requires-python = ">=3.10"

classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dynamic = [ "version" ]

dependencies = [
  # FIXME: the fsi should not depend from DropBox (dropbox should be optional)
  "dropbox>=10.10; platform_system!='iOS'",
  "packaging>=25; platform_system!='iOS'",
  "paramiko>=4; platform_system!='iOS'",
  "pyaes-whl>=1.6.1",
  "pycryptodome>=3.23; platform_system!='iOS'",
  "pyftpdlib>=2.0.1",
  "pyperclip>=1.9; platform_system!='iOS'",
  "pyte>=0.8.2",
  "pytz>=2025.2; platform_system!='iOS'",
  "requests>=2.32.4",
  "rsa>=4.9.1",
  "standard-telnetlib>=3.13",
]

optional-dependencies.testing = [
  "flake8>=3.7.9",
  "pyparsing",
  "pytest",
  "ruff>=0.12.7",
]
urls.Homepage = "https://github.com/ywangd/stash/"
urls.Issues = "https://github.com/ywangd/stash/issues"
urls.Repository = "https://github.com/ywangd/stash/"

# cli entry point
scripts.stash = "stash.__main__:main"

[dependency-groups]
dev = [
  "flake8>=3.7.9",
  "pyparsing",
  "pytest",
  "ruff>=0.12.7",
  "types-pytz>=2025.2.0.20250809",
]

[tool.setuptools]
#zip-safe = false
py-modules = [
  "core",
  "getstash",
  "launch_stash",
  "__init__",
  "__main__",
]

[tool.setuptools.dynamic]
version = { attr = "core.__version__" }

[tool.setuptools.packages.find]
where = [ "." ]
include = [
  "*",
]
exclude = [
  ".idea",
  "*venv",
  "build",
  "dist",
  "*.egg-info",
  "__pycache__",
]

[tool.setuptools.package-data]
"*" = [ "**/*" ]

[tool.ruff]
target-version = "py310"
lint.ignore = [
  "E402",
  "E721",
  "E722",
  "E731",
  "F401",
  "F841",
]

[tool.pytest.ini_options]
# The testpaths option (optional, but good practice to specify where tests are)
testpaths = [ "tests/" ]

# Options that take a list of values
addopts = [
  "-v",
  "--showlocals",
  "--verbose",
  "--show-capture=all",
  "--log-level=debug",
  "--ignore=tests/system/data/",
  "--ignore=tests/pip/test_pip.py",
]
